+++ test-suite/Main.hs
@@ -613,6 +613,7 @@ main = do
       , golden "bigint-factorial-py" "bigint-factorial-py"
       , golden "bigint-overflow-cpp" "bigint-overflow-cpp"
       , golden "bigint-overflow-r" "bigint-overflow-r"
+      , golden "int-overflow" "int-overflow"
       , -- unary minus / negate operator
         golden "negate-unary-operator" "negate-unary-operator"
       ]